♻️ group component by categories, move project images into a new folder
