+++ components/layout/ContainerBlock.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Head from "next/head";
 import { useRouter } from "next/router";
-import Navbar from "./Navbar";
+import Navbar from "../navigation/Navbar";
 import Footer from "./Footer";
 
 export default function ContainerBlock({ children, ...customMeta }) {
