+++ pages/500.js
@@ -1,5 +1,5 @@
-import ContainerBlock from "../components/ContainerBlock";
-import Component500 from "@components/Error500";
+import ContainerBlock from "@components/layout/ContainerBlock";
+import Component500 from "@components/commons/Error500";
 
 export default function Custom404() {
   return (
