+++ pages/experience.js
@@ -1,5 +1,5 @@
-import ContainerBlock from "../components/ContainerBlock";
-import Experience from "../components/Experience";
+import ContainerBlock from "@components/layout/ContainerBlock";
+import Experience from "@components/experience/Experience";
 
 export default function experience() {
   return (
