+++ pages/index.js
@@ -1,5 +1,5 @@
-import ContainerBlock from "../components/ContainerBlock";
-import LatestCode from "../components/LatestCode";
+import ContainerBlock from "@components/layout/ContainerBlock";
+import LatestCode from "@components/projects/LatestCode";
 import Hero from "../components/Hero";
 import getLatestRepos from "@lib/getLatestRepos";
 import userData from "@constants/data";
